Add 'Synchronize JBang' command
The `Synchronize JBang` command is available when right-clicking on a JBang script.
If the script belongs to a non-Java project, the Java nature will be added to that project.
The source folder containing the JBang file will automatically be added to the project's classpath.
The `JBang Dependencies` classpath container will be added to the classpath as well.

Caveats:
- JBang files and dependencies will leak to the project's main and test-scoped sources
- the project's main files and dependencies will leak to the JBang script
- If the script file is at the root of the project and there are other source folders nested beneath that root, compilation errors will occur in those folders

Signed-off-by: Fred Bricon <fbricon@gmail.com>